+++ magpie/api/login/__init__.py
@@ -1,3 +1,4 @@
+from magpie.constants import get_constant
 from magpie.utils import get_logger
 
 LOGGER = get_logger(__name__)
@@ -11,4 +12,7 @@ def includeme(config):
     config.add_route(**s.service_api_route_info(s.SigninAPI))
     config.add_route(**s.service_api_route_info(s.ProvidersAPI))
     config.add_route(**s.service_api_route_info(s.ProviderSigninAPI))
+    if get_constant("MAGPIE_NETWORK_MODE", settings_name="magpie.network_mode"):
+        config.add_route(**s.service_api_route_info(s.TokenAPI))
+        config.add_route(**s.service_api_route_info(s.TokenValidateAPI))
     config.scan()