+++ magpie/api/login/login.py
@@ -1,6 +1,10 @@
 import json
+import uuid
+from datetime import datetime, timedelta
 from typing import TYPE_CHECKING
 
+import jwt
+import requests
 from authomatic.adapters import WebObAdapter
 from authomatic.core import Credentials, LoginResult, resolve_provider_class
 from authomatic.exceptions import OAuth2Error
@@ -34,7 +38,7 @@ from magpie.api import requests as ar
 from magpie.api import schemas as s
 from magpie.api.management.user.user_formats import format_user
 from magpie.api.management.user.user_utils import create_user
-from magpie.constants import get_constant
+from magpie.constants import get_constant, protected_user_name_regex
 from magpie.security import authomatic_setup, get_providers
 from magpie.utils import (
     CONTENT_TYPE_JSON,
@@ -47,14 +51,23 @@ from magpie.utils import (
 
 if TYPE_CHECKING:
     from magpie.typedefs import Session, Str
+    from typing import Optional
 
 LOGGER = get_logger(__name__)
 
+MAGPIE_NETWORK_MODE = get_constant("MAGPIE_NETWORK_MODE", settings_name="magpie.network_mode")
 
 # dictionaries of {'provider_id': 'provider_display_name'}
 MAGPIE_DEFAULT_PROVIDER = get_constant("MAGPIE_DEFAULT_PROVIDER")
 MAGPIE_INTERNAL_PROVIDERS = {MAGPIE_DEFAULT_PROVIDER: MAGPIE_DEFAULT_PROVIDER.capitalize()}
 MAGPIE_EXTERNAL_PROVIDERS = get_providers()
+
+if MAGPIE_NETWORK_MODE:
+    MAGPIE_NETWORK_TOKEN_NAME = get_constant("MAGPIE_NETWORK_TOKEN_NAME")
+    MAGPIE_NETWORK_PROVIDER = get_constant("MAGPIE_NETWORK_PROVIDER")
+    MAGPIE_INSTANCE_NAME = get_constant("MAGPIE_INSTANCE_NAME")
+    MAGPIE_EXTERNAL_PROVIDERS[MAGPIE_NETWORK_PROVIDER] = MAGPIE_NETWORK_PROVIDER
+
 MAGPIE_PROVIDER_KEYS = frozenset(set(MAGPIE_INTERNAL_PROVIDERS) | set(MAGPIE_EXTERNAL_PROVIDERS))
 
 
@@ -130,8 +143,8 @@ def sign_in_view(request):
     # request handler. Catch that specific exception and return it to bypass the EXCVIEW tween that result in that
     # automatic convert to return 403 directly.
     try:
-        anonymous = get_constant("MAGPIE_ANONYMOUS_USER", request)
-        ax.verify_param(user_name, not_equal=True, param_compare=anonymous, param_name="user_name",
+        anonymous_regex = protected_user_name_regex(include_admin=False, settings_container=request)
+        ax.verify_param(user_name, not_matches=True, param_compare=anonymous_regex, param_name="user_name",
                         http_error=HTTPForbidden, msg_on_fail=s.Signin_POST_ForbiddenResponseSchema.description)
     except HTTPForbidden as http_error:
         return http_error
@@ -239,17 +252,12 @@ def new_user_external(external_user_name, external_id, email, provider_name, db_
     return user
 
 
-def login_success_external(request, external_user_name, external_id, email, provider_name):
-    # type: (Request, Str, Str, Str, Str) -> HTTPException
+def login_success_external(request, user):
+    # type: (Request, models.User) -> HTTPException
     """
     Generates the login response in case of successful external provider identification.
     """
-    # find possibly already registered user by external_id/provider
-    user = ExternalIdentityService.user_by_external_id_and_provider(external_id, provider_name, request.db)
-    if user is None:
-        # create new user with an External Identity
-        user = new_user_external(external_user_name=external_user_name, external_id=external_id,
-                                 email=email, provider_name=provider_name, db_session=request.db)
+
     # set a header to remember user (set-cookie)
     headers = remember(request, user.id)
 
@@ -271,10 +279,27 @@ def login_success_external(request, external_user_name, external_id, email, prov
                          http_kwargs={"location": homepage_route, "headers": headers})
 
 
+def user_from_token(request, token):
+    # type: (Request, Str) -> Optional[models.User]
+    """
+    Return the ``User`` associated with the token as long as the token is valid and issued by this Magpie instance.
+    """
+    magpie_secret = get_constant("MAGPIE_SECRET", request, settings_name="magpie.secret")
+    decoded_token = jwt.decode(token, magpie_secret, algorithms="HS256", issuer=MAGPIE_INSTANCE_NAME,
+                               options={"require": ["exp", "token"]})
+
+    network_token = (request.db.query(models.NetworkToken)
+                     .join(models.User)
+                     .filter(models.NetworkToken.token == decoded_token["token"])
+                     .first())
+    if network_token:
+        return network_token.user
+
+
 @s.ProviderSigninAPI.get(schema=s.ProviderSignin_GET_RequestSchema, tags=[s.SessionTag],
                          response_schemas=s.ProviderSignin_GET_responses)
 @view_config(route_name=s.ProviderSigninAPI.name, permission=NO_PERMISSION_REQUIRED)
-def authomatic_login_view(request):
+def external_login_view(request):
     """
     Signs in a user session using an external provider.
     """
@@ -282,58 +307,99 @@ def authomatic_login_view(request):
     response = Response()
     verify_provider(provider_name)
     try:
-        authomatic_handler = authomatic_setup(request)
+        if provider_name == MAGPIE_NETWORK_PROVIDER:
+            if "Authorization" in request.headers:
+                token_type, token = request.headers.get("Authorization").split()
+                if token_type != "Bearer":
+                    ax.raise_http(http_error=HTTPBadRequest,
+                                  detail=s.ProviderSignin_GET_BadRequestResponseSchema.description)
+                authenticated_user = None
+                try:
+                    authenticated_user = user_from_token(request, token)
+                except jwt.exceptions.InvalidIssuerError:
+                    decoded_token = jwt.decode(token, options={"verify_signature": False})
+                    issuer = decoded_token.get("iss")
+                    node = request.db.query(models.NetworkNode).filter(models.NetworkNode.name == issuer).first()
+                    if node:
+                        response = requests.get("{}{}".format(node.url, s.TokenValidateAPI.path),
+                                                headers={"Accept": CONTENT_TYPE_JSON})
 
-        # if we directly have the Authorization header, bypass authomatic login and retrieve 'userinfo' to signin
-        if "Authorization" in request.headers and "authomatic" not in request.cookies:
-            provider_config = authomatic_handler.config.get(provider_name, {})
-            provider_class = resolve_provider_class(provider_config.get("class_"))
-            provider = provider_class(authomatic_handler, adapter=None, provider_name=provider_name)
-            # provide the token user data, let the external provider update it on login afterwards
-            token_type, access_token = request.headers.get("Authorization").split()
-            data = {"access_token": access_token, "token_type": token_type}
-            cred = Credentials(authomatic_handler.config, token=access_token, token_type=token_type, provider=provider)
-            provider.credentials = cred
-            result = LoginResult(provider)
-            # pylint: disable=W0212
-            result.provider.user = result.provider._update_or_create_user(data, credentials=cred)  # noqa: W0212
-
-        # otherwise, use the standard login procedure
-        else:
-            result = authomatic_handler.login(WebObAdapter(request, response), provider_name)
-            if result is None:
-                if response.location is not None:
-                    return HTTPTemporaryRedirect(location=response.location, headers=response.headers)
-                return response
-
-        if result:
-            if result.error:
-                # Login procedure finished with an error.
-                error = result.error.to_dict() if hasattr(result.error, "to_dict") else result.error
-                LOGGER.debug("Login failure with error. [%r]", error)
-                return login_failure_view(request, reason=result.error.message)
-            if result.user:
-                # OAuth 2.0 and OAuth 1.0a provide only limited user data on login,
-                # update the user to get more info.
-                if not (result.user.name and result.user.id):
-                    try:
-                        response = result.user.update()
-                    # this error can happen if providing incorrectly formed authorization header
-                    except OAuth2Error as exc:
-                        LOGGER.debug("Login failure with Authorization header.")
-                        ax.raise_http(http_error=HTTPBadRequest, content={"reason": str(exc.message)},
-                                      detail=s.ProviderSignin_GET_BadRequestResponseSchema.description)
-                    # verify that the update procedure succeeded with provided token
-                    if 400 <= response.status < 500:
-                        LOGGER.debug("Login failure with invalid token.")
+                        if response.status_code != HTTPOk.code:
+                            ax.raise_http(http_error=HTTPUnauthorized,
+                                          detail=s.ProviderSignin_GET_UnauthorizedTokenSchema.description)
+                        user_name = response.json().get("user_name")
+                        authenticated_user = (node.users.filter(models.User.name == user_name).first() or
+                                              node.anonymous_user)
+                    else:
                         ax.raise_http(http_error=HTTPUnauthorized,
-                                      detail=s.ProviderSignin_GET_UnauthorizedResponseSchema.description)
-                # create/retrieve the user using found details from login provider
-                return login_success_external(request,
-                                              external_id=result.user.username or result.user.id,
-                                              email=result.user.email,
-                                              provider_name=result.provider.name,
-                                              external_user_name=result.user.name)
+                                      detail=s.ProviderSignin_GET_UnauthorizedTokenSchema.description)
+                except jwt.exceptions.PyJWTError as exc:
+                    ax.raise_http(http_error=HTTPUnauthorized, content={"reason": str(exc)},
+                                  detail=s.ProviderSignin_GET_UnauthorizedTokenSchema.description)
+                if authenticated_user:
+                    return login_success_external(request, authenticated_user)
+            else:
+                ax.raise_http(http_error=HTTPBadRequest,
+                              detail=s.ProviderSignin_GET_BadRequestResponseSchema.description)
+        else:
+            authomatic_handler = authomatic_setup(request)
+
+            # if we directly have the Authorization header, bypass authomatic login and retrieve 'userinfo' to signin
+            if "Authorization" in request.headers and "authomatic" not in request.cookies:
+                provider_config = authomatic_handler.config.get(provider_name, {})
+                provider_class = resolve_provider_class(provider_config.get("class_"))
+                provider = provider_class(authomatic_handler, adapter=None, provider_name=provider_name)
+                # provide the token user data, let the external provider update it on login afterwards
+                token_type, access_token = request.headers.get("Authorization").split()
+                data = {"access_token": access_token, "token_type": token_type}
+                cred = Credentials(authomatic_handler.config, token=access_token, token_type=token_type,
+                                   provider=provider)
+                provider.credentials = cred
+                result = LoginResult(provider)
+                # pylint: disable=W0212
+                result.provider.user = result.provider._update_or_create_user(data, credentials=cred)  # noqa: W0212
+
+            # otherwise, use the standard login procedure
+            else:
+                result = authomatic_handler.login(WebObAdapter(request, response), provider_name)
+                if result is None:
+                    if response.location is not None:
+                        return HTTPTemporaryRedirect(location=response.location, headers=response.headers)
+                    return response
+
+            if result:
+                if result.error:
+                    # Login procedure finished with an error.
+                    error = result.error.to_dict() if hasattr(result.error, "to_dict") else result.error
+                    LOGGER.debug("Login failure with error. [%r]", error)
+                    return login_failure_view(request, reason=result.error.message)
+                if result.user:
+                    # OAuth 2.0 and OAuth 1.0a provide only limited user data on login,
+                    # update the user to get more info.
+                    if not (result.user.name and result.user.id):
+                        try:
+                            response = result.user.update()
+                        # this error can happen if providing incorrectly formed authorization header
+                        except OAuth2Error as exc:
+                            LOGGER.debug("Login failure with Authorization header.")
+                            ax.raise_http(http_error=HTTPBadRequest, content={"reason": str(exc.message)},
+                                          detail=s.ProviderSignin_GET_BadRequestResponseSchema.description)
+                        # verify that the update procedure succeeded with provided token
+                        if 400 <= response.status < 500:
+                            LOGGER.debug("Login failure with invalid token.")
+                            ax.raise_http(http_error=HTTPUnauthorized,
+                                          detail=s.ProviderSignin_GET_UnauthorizedResponseSchema.description)
+                    # create/retrieve the user using found details from login provider
+                    # find possibly already registered user by external_id/provider
+                    external_id = result.user.username or result.user.id
+                    user = ExternalIdentityService.user_by_external_id_and_provider(external_id, provider_name,
+                                                                                    request.db)
+                    if user is None:
+                        # create new user with an External Identity
+                        user = new_user_external(external_user_name=result.user.name, external_id=external_id,
+                                                 email=result.user.email, provider_name=result.provider.name,
+                                                 db_session=request.db)
+                    return login_success_external(request, user)
     except Exception as exc:
         exc_msg = "Unhandled error during external provider '{}' login. [{!s}]".format(provider_name, exc)
         LOGGER.exception(exc_msg, exc_info=True)
@@ -371,6 +437,71 @@ def get_session_view(request):
     session_json = ax.evaluate_call(lambda: _get_session(request), http_error=HTTPInternalServerError,
                                     msg_on_fail=s.Session_GET_InternalServerErrorResponseSchema.description)
     return ax.valid_http(http_success=HTTPOk, detail=s.Session_GET_OkResponseSchema.description, content=session_json)
+
+
+if MAGPIE_NETWORK_MODE:
+    @s.TokenAPI.patch(schema=s.Token_PATCH_RequestBodySchema, tags=[s.SessionTag],
+                      response_schemas=s.Token_PATCH_responses)
+    @view_config(route_name=s.TokenAPI.name, request_method="PATCH", permission=Authenticated)
+    def token_view(request):
+        """
+        Get a token. Generates a new token every time this route is accessed so this can also
+        be used to invalidate a previously generated token.
+        """
+        magpie_secret = get_constant("MAGPIE_SECRET", request, settings_name="magpie.secret")
+        default_expiry = get_constant("MAGPIE_DEFAULT_TOKEN_EXPIRY", request,
+                                      settings_name="magpie.default_token_expiry")
+        magpie_instance_name = get_constant("MAGPIE_INSTANCE_NAME",
+                                            request,
+                                            settings_name="magpie.instance_name")
+
+        expires = request.GET.get("expires", default_expiry)
+        try:
+            expires = int(expires)
+        except ValueError:
+            expires = int(default_expiry)
+
+        expiry = datetime.utcnow() + timedelta(seconds=expires)
+
+        def upsert_token():
+            network_token = request.db.query(models.NetworkToken).filter(
+                models.NetworkToken.user_id == request.user.id).first()
+            if network_token:
+                network_token.token = uuid.uuid4()
+            else:
+                network_token = models.NetworkToken(user_id=request.user.id)
+                request.db.add(network_token)
+            return network_token.token
+
+        token_value = ax.evaluate_call(lambda: upsert_token(), fallback=lambda: request.db.rollback(),
+                                       http_error=HTTPInternalServerError)
+
+        token = jwt.encode({"token": str(token_value),
+                            "exp": expiry,
+                            "iss": magpie_instance_name}, magpie_secret, algorithm="HS256")
+
+        return ax.valid_http(http_success=HTTPOk, detail=s.Token_PATCH_OkResponseSchema.description,
+                             content={"token": token})
+
+
+    @s.TokenValidateAPI.get(schema=s.TokenValidate_GET_RequestBodySchema, tags=[s.SessionTag],
+                            response_schemas=s.TokenValidate_GET_responses)
+    @view_config(route_name=s.TokenValidateAPI.name, permission=NO_PERMISSION_REQUIRED)
+    def validate_token_view(request):
+        """ Validate a token """
+        token = request.GET.get("token")
+        try:
+            authenticated_user = user_from_token(request, token)
+            if authenticated_user:
+                return ax.valid_http(http_success=HTTPOk,
+                                     detail=s.TokenValidate_GET_OkResponseSchema.description,
+                                     content={"user_name": authenticated_user.user_name})
+            else:
+                ax.raise_http(http_error=HTTPUnauthorized,
+                              detail=s.TokenValidate_GET_BadRequestResponseSchema.description)
+        except jwt.exceptions.PyJWTError as exc:
+            ax.raise_http(http_error=HTTPUnauthorized, content={"reason": str(exc)},
+                          detail=s.TokenValidate_GET_BadRequestResponseSchema.description)
 
 
 @s.ProvidersAPI.get(tags=[s.SessionTag], response_schemas=s.Providers_GET_responses)